demo_project/requirements.txt
asgiref==3.4.1
certifi==2022.9.24
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.16
django-cors-headers==3.10.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
joblib==1.1.1
MarkupSafe==2.0.1
numpy==1.19.5
packaging==21.3
PyJWT==1.7.1
pyparsing==3.0.9
pytz==2022.5
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==0.24.2
scipy==1.5.4
sqlparse==0.4.3
threadpoolctl==3.1.0
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.12
